fix:바텀바 core폴더로 이동
